+++ lang/test/org/partiql/lang/planner/transforms/AstToLogicalVisitorTransformTests.kt
@@ -237,10 +237,39 @@ class AstToLogicalVisitorTransformTests {
                     )
                 }
             ),
+            TestCase(
+                "PIVOT x.v AT x.a FROM << {'a': 'first', 'v': 'john'}, {'a': 'last', 'v': 'doe'} >> as x",
+                PartiqlLogical.build {
+                    query(
+                        pivot(
+                            input = scan(
+                                bag(
+                                    struct(
+                                        listOf(
+                                            structField(lit(ionString("a")), lit(ionString("first"))),
+                                            structField(lit(ionString("v")), lit(ionString("john"))),
+                                        )
+                                    ),
+                                    struct(
+                                        listOf(
+                                            structField(lit(ionString("a")), lit(ionString("last"))),
+                                            structField(lit(ionString("v")), lit(ionString("doe"))),
+                                        )
+                                    )
+                                ),
+                                asDecl = varDecl("x"),
+                            ),
+                            key = path(id("x"), listOf(pathExpr(lit(ionString("v"))))),
+                            value = path(id("x"), listOf(pathExpr(lit(ionString("a"))))),
+                        )
+                    )
+                }
+            )
         )
     }
 
     data class ProblemTestCase(val sql: String, val expectedProblem: Problem)
+
     @ParameterizedTest
     @ArgumentsSource(ArgumentsForProblemTests::class)
     fun `unimplemented feautres are blocked`(tc: ProblemTestCase) {
@@ -267,7 +296,6 @@ class AstToLogicalVisitorTransformTests {
             // SELECT queries are not implemented
             ProblemTestCase("SELECT b.* FROM bar AS b GROUP BY a", unimplementedProblem("GROUP BY", 1, 26)),
             ProblemTestCase("SELECT b.* FROM bar AS b HAVING x", unimplementedProblem("HAVING", 1, 33)),
-            ProblemTestCase("PIVOT v AT n FROM data AS d", unimplementedProblem("PIVOT", 1, 1)),
 
             // DDL is  not implemented
             ProblemTestCase("CREATE TABLE foo", unimplementedProblem("CREATE TABLE", 1, 1)),